+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+0.9.0 (xxxx-xx-xx)
+==================
+  * Added supporting Django v3.0
+  * Added supporting render http-server
+  * Added `requests` into `install_requires`
+  * Added Python 3.8 in tests
+  * Added MJML 4.5.1 in tests
+  * Upgraded MJML to 4.5.1 in dockerfile
+  * Upgraded Node to v12 for tcp-server
+  * Reorganized tests
+
+
 0.8.0 (2019-07-29)
 ==================
   * Fixed a trouble with unicode